+++ src/store/AuthStore.ts
@@ -3,23 +3,16 @@ import {
   types,
   flow,
   cast,
-  SnapshotOut,
-  SnapshotIn,
   getParent,
+  getSnapshot,
 } from 'mobx-state-tree';
 import api from '../services/api/Api';
 import Storage from '../services/storage';
 import { ANALYTICS_EVENT, sendAnalyticsEvent } from '../utils/analytics';
 
+import { UserModel } from './models';
+
 const States = ['IDLE' as const, 'PROCESSING' as const, 'ERROR' as const];
-
-const UserModel = types.model({
-  id: types.number,
-});
-
-export interface IUserModel extends Instance<typeof UserModel> {}
-export interface User extends SnapshotOut<typeof UserModel> {}
-export interface UserIn extends SnapshotIn<typeof UserModel> {}
 
 export const AuthStore = types
   .model({
@@ -38,6 +31,9 @@ export const AuthStore = types
     },
     get showLoginModal() {
       return self.isLoginModalOpen && !this.isLoggedIn;
+    },
+    get userData() {
+      return self.user ? getSnapshot(self.user) : undefined;
     },
   }))
   .actions(self => {
@@ -171,6 +167,40 @@ export const AuthStore = types
       }
     });
 
+    const getMe = flow(function* (params: API.GetMe = {}) {
+      self.state = 'PROCESSING';
+
+      const response: API.GeneralResponse<API.RES.GetMe> = yield api.getMe(
+        params
+      );
+
+      if (response.kind === 'ok') {
+        self.user = cast(response.data);
+        self.state = 'IDLE';
+        return { success: true };
+      } else {
+        self.state = 'ERROR';
+        return { success: false, error: response.data };
+      }
+    });
+
+    const setUserTags = flow(function* (params: API.SetUserTags) {
+      self.state = 'PROCESSING';
+
+      const response: API.GeneralResponse<API.RES.SetUserTags> =
+        yield api.setUserTags(params);
+
+      if (response.kind === 'ok') {
+        self.user = cast(response.data);
+
+        self.state = 'IDLE';
+        return { success: true };
+      } else {
+        self.state = 'ERROR';
+        return { success: false };
+      }
+    });
+
     return {
       register,
       openLoginModal,
@@ -181,6 +211,8 @@ export const AuthStore = types
       resetPassword,
       logout,
       deleteAccount,
+      getMe,
+      setUserTags,
     };
   });